fix deprecation warnings and catch OptimizeWarning (to work with new numpy)
This slightly changes the outcome: when fitting an exponential function,
set mu to nan if OptimizeWarning is raised.
The other changes make sure index arrays have the length of the arrays
they are used with (this should not change the outcome).
Additionally removed a call that accidentally turned warnings globally
into errors.